feat: runtime index prefix (#24) + typed source→sink value contract

+++ libs/1-sinks/2-opensearch/benches/opensearch.rs
@@ -129,7 +129,7 @@ fn mapping() -> IndexMapping {
 /// A realistic flat document for id `i`, matching the mapping above.
 fn document(i: usize) -> GenericValue {
     let mut map = BTreeMap::new();
-    map.insert("id".to_owned(), GenericValue::Int(i as i64));
+    map.insert("id".to_owned(), GenericValue::BigInt(i as i64));
     map.insert(
         "email".to_owned(),
         GenericValue::String(format!("user{i}@example.com")),
